optimization of code, unstable!

+++ include/algorithms.h
@@ -10,7 +10,7 @@ namespace RALG {
 void realgen_swap(float &a, float &b);
 int partition(float arr[], int l, int r);
 float kthSmallest_(float arr[], int l, int r, int k);
-int kthSmallest(vector<float> &fitnessValues, int l, int r, int k);
+float kthSmallest(vector<float> &fitnessValues, int l, int r, int k);
 int argKthSmallest(vector<float> &fitnessValues, int l, int r, int k);
 }
